refactoring: app initialization in App component. Removed unused AppInitializer.tsx and AppLoaderLayout.tsx
