--- src/app/index.ts
@@ -1,2 +1,0 @@
-export { Providers } from './providers';
-export { AppInitializer } from './initialization';
